requirements.txt
torch
torchvision
pillow
pandas
streamlit
numpy